// src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        StudentRead: {
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /**
             * Personal Id
             * Format: uuid
             */
            personal_id: string;
            /** Name */
            name: string;
            /** Email */
            email: string;
            /** Phone */
            phone: string | null;
            /** Is Active */
            is_active: boolean;
            /** Monthly Fee */
            monthly_fee: string | null;
            /** Due Day */
            due_day: number | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}